# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Дом в порядке API
  version: 0.1.0
  description: Собственный API MVP. Все маршруты `/api/*`, кроме health, требуют cookie-сессию диспетчера.
servers:
  - url: https://jorwey.ru
    description: Production
paths:
...
  /api/incidents:
    get:
      security: [{ cookieAuth: [] }]
      summary: Очередь или архив инцидентов
      parameters:
        - { name: archive, in: query, schema: { type: boolean } }
        - name: sort
          in: query
          description: Порядок очереди по дате создания или числу подключившихся жителей.
          schema: { type: string, enum: [importance_desc, importance_asc, date_desc, date_asc], default: importance_desc }
      responses: { '200': { description: Список, content: { application/json: { schema: { $ref: '#/components/schemas/IncidentList' } } } }, '401': { $ref: '#/components/responses/Unauthorized' } }
  /api/incidents/{id}:
    get:
      security: [{ cookieAuth: [] }]
      summary: Карточка инцидента
      parameters: [{ $ref: '#/components/parameters/IncidentId' }]
      responses: { '200': { description: Карточка }, '404': { $ref: '#/components/responses/NotFound' } }
    patch:
      security: [{ cookieAuth: [] }]
      summary: Смена статуса, срока или комментария
      parameters: [{ $ref: '#/components/parameters/IncidentId' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/IncidentPatch' } } } }
      responses: { '200': { description: Инцидент изменён }, '400': { description: Ошибка DTO }, '409': { description: Конфликт версии или статуса } }
  /api/incidents/{id}/comments/{commentId}:
    delete:
      security: [{ cookieAuth: [] }]
      summary: Удалить комментарий жителя
      parameters:
        - { $ref: '#/components/parameters/IncidentId' }
        - { name: commentId, in: path, required: true, schema: { type: string, format: uuid } }
      responses: { '204': { description: Комментарий удалён }, '404': { $ref: '#/components/responses/NotFound' } }
...
  /api/settings/uk-code:
    get:
      security: [{ cookieAuth: [] }]
      summary: Маскированный код УК или раскрытие по ?reveal=true
      description: Раскрытие полного кода требует роли supervisor и записывается в аудит.
      parameters: [{ name: reveal, in: query, schema: { type: boolean, default: false } }]
      responses: { '200': { description: no-store ответ с кодом или маской }, '403': { $ref: '#/components/responses/Forbidden' } }
    patch:
      security: [{ cookieAuth: [] }]
      summary: Сменить общий десятизначный код УК
      description: Требуется роль supervisor. Операция записывается в административный аудит.
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [code], properties: { code: { type: string, pattern: '^\\d{10}$' } } } } } }
      responses: { '200': { description: Код обновлён }, '400': { description: Неверный формат }, '403': { $ref: '#/components/responses/Forbidden' } }
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: smart_home_admin }
  parameters:
    IncidentId: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
  responses:
    Unauthorized: { description: Требуется сессия, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Forbidden: { description: Недостаточно прав, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    NotFound: { description: Не найдено, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  schemas:
...
    Error: { type: object, required: [code], properties: { code: { type: string }, requestId: { type: string } } }
...
    IncidentList: { type: object, required: [items], properties: { items: { type: array, items: { $ref: '#/components/schemas/Incident' } } } }
    Incident: { type: object, required: [id, title, status, version], properties: { id: { type: string, format: uuid }, title: { type: string }, status: { type: string, enum: [new, acknowledged, in_progress, resolved, rejected] }, version: { type: integer, minimum: 1 } } }
    IncidentPatch: { type: object, required: [version, dueAt], properties: { version: { type: integer, minimum: 1 }, status: { type: string, enum: [new, acknowledged, in_progress, resolved, rejected] }, dueAt: { type: [string, 'null'], format: date-time }, publicComment: { type: string, minLength: 3, maxLength: 1000 } } }
...